removed the second redirect folder and added new redirecting and updated books to the original tests
